* commit '981a208b4488f679a96d24226d139982fa3f0193 [formerly fbb07f350990f9a2b5b732961e64d7d94c5f63b3]':
  Allow disbursement creation if grantee has disbursements
  Changes to amend grant report generation date and report and disbursement switch
  Disbursement rows fix for grantees
  Multiple grants in published state for grantee scenario
  Fix for delete reports
  Fix for close state to assignment nulls
  Get disbursed data for all amended grants
  Soft delete bug fix
  Amended grants email messaging


Former-commit-id: 2b33e320feadd0f29b3f950dfef875af09e1a795

+++ src/main/java/org/codealpha/gmsservice/services/ReportService.java
@@ -520,7 +520,13 @@ public class ReportService {
             }
         }
 
-        String message = msgConfigValue.replaceAll("%GRANT_NAME%", finalReport.getGrant().getName())
+        String grantName = "";
+        if(finalReport.getGrant().getReferenceNo()!=null){
+            grantName = "[".concat(finalReport.getGrant().getReferenceNo()).concat("] ").concat(finalReport.getGrant().getName());
+        }else{
+            grantName = finalReport.getGrant().getName();
+        }
+        String message = msgConfigValue.replaceAll("%GRANT_NAME%", grantName)
                 .replaceAll("%REPORT_NAME%", finalReport.getName()).replaceAll("%REPORT_LINK%", granteeUrl)
                 .replaceAll("%CURRENT_STATE%", currentState).replaceAll("%CURRENT_OWNER%", currentOwner)
                 .replaceAll("%PREVIOUS_STATE%", previousState).replaceAll("%PREVIOUS_OWNER%", previousOwner)
@@ -537,7 +543,7 @@ public class ReportService {
                 .replaceAll("%GRANTER%", finalReport.getGrant().getGrantorOrganization().getName())
                 .replaceAll("%ENTITY_TYPE%", "report")
                 .replaceAll("%PREVIOUS_ASSIGNMENTS%", getAssignmentsTable(previousApprover, newApprover))
-                .replaceAll("%ENTITY_NAME%", finalReport.getName() + " of grant " + finalReport.getGrant().getName());
+                .replaceAll("%ENTITY_NAME%", finalReport.getName() + " of grant " + grantName);
         String subject = subConfigValue.replaceAll("%REPORT_NAME%", finalReport.getName());
 
         return new String[] { subject, message };
@@ -755,7 +761,7 @@ public class ReportService {
     }
 
     public List<Report> getReportsForGrant(Grant grant) {
-        return reportRepository.getReportsByGrant(grant);
+        return reportRepository.getReportsByGrant(grant.getId());
     }
 
     public void setAssignmentHistory(ReportAssignmentsVO assignmentsVO) {
@@ -786,21 +792,27 @@ public class ReportService {
     }
 
     public void deleteReport(Report report) {
-        for (ReportSpecificSection section : getReportSections(report)) {
-            List<ReportSpecificSectionAttribute> attribs = getSpecificSectionAttributesBySection(section);
-            for (ReportSpecificSectionAttribute attribute : attribs) {
-                List<ReportStringAttribute> strAttribs = getReportStringAttributesByAttribute(attribute);
-                deleteStringAttributes(strAttribs);
+
+        if(checkIfReportMovedThroughWFAtleastOnce(report.getId())){
+            report.setDeleted(true);
+            saveReport(report);
+        }else {
+            for (ReportSpecificSection section : getReportSections(report)) {
+                List<ReportSpecificSectionAttribute> attribs = getSpecificSectionAttributesBySection(section);
+                for (ReportSpecificSectionAttribute attribute : attribs) {
+                    List<ReportStringAttribute> strAttribs = getReportStringAttributesByAttribute(attribute);
+                    deleteStringAttributes(strAttribs);
+                }
+                deleteSectionAttributes(attribs);
+                deleteSection(section);
             }
-            deleteSectionAttributes(attribs);
-            deleteSection(section);
-        }
 
-        reportRepository.delete(report);
+            reportRepository.delete(report);
 
-        GranterReportTemplate template = granterReportTemplateRepository.findById(report.getTemplate().getId()).get();
-        if (!template.isPublished()) {
-            deleteReportTemplate(template);
+            GranterReportTemplate template = granterReportTemplateRepository.findById(report.getTemplate().getId()).get();
+            if (!template.isPublished()) {
+                deleteReportTemplate(template);
+            }
         }
     }
